move functions to autoload script
apparently the tplugin plugin manager can do some lazy loading by
parsing the plugin files for commands, invoking their completion
functions, etc, but that requires that the completion plugin is public
(and perhaps in an autoload script?). So, since we are going the
autoload route for the command completion function, might as well move
all the ag functions into an autoload script.

closes #7